+++ readme_magic/__init__.py
@@ -1,6 +1,6 @@
 ﻿"""ReadmeMagic - One spell, beautiful README"""
 
-__version__ = "2.1.0"
+__version__ = "2.1.1"
 __author__ = "GetIT-Sunday"
 __description__ = "Create professional, beautiful, high-converting README.md files"